chore(core): dedicated expression module
There is now a more pronounced difference between expressions and transforms, so it's helpful to split the former out into a dedicated module. Previously the only transform implemented had been calculate(), which in turn always ran simple string expression interpolation. That's no longer the case now that more transforms have been added, so mashing the expression evaluation into the transform module seems more obviously out of place.
